requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
djangorestframework
markdown==3.3.4
django-filter==2.4.0
uritemplate==3.0.1
PyYAML==5.4.1
Pygments==2.8.1
Pillow==8.1.2
drf-spectacular
# dj-rest-auth
# djangorestframework-simplejwt